Only get time from RTC if the difference is not too large.
The periodic sync is used to fix time drift between RTC and the internal
clock, so the difference should never be too large.

For some reason (maybe dodgy connections due to the cramped case?) we receive
a value from the RTC that is wildly off -- multiple hours and minutes. These should
be skipped.